# itsm-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: AI-Native ITSM 平台 API：ITIL v4 核心流程（工单/事件/问题/变更/发布/服务请求）、CMDB、BPMN 工作流、知识库
    RAG、SLA、RBAC 与多租户隔离。
  title: ITSM API
  version: 1.6.9
paths:
...
  /api/v1/departments:
    get:
      description: 返回当前租户的平铺部门列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取部门列表
      tags:
      - 组织管理
  /api/v1/departments/tree:
    get:
      description: 返回当前租户的部门层级树
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取部门树
      tags:
      - 组织管理
...
  /api/v1/org/departments:
    post:
      consumes:
      - application/json
      description: 在当前租户下创建部门（name、code 必填）
      parameters:
      - description: 部门信息（name、code、description、managerId、parentId）
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建部门
      tags:
      - 组织管理
  /api/v1/org/departments/{id}:
    delete:
      description: 删除当前租户下的部门
      parameters:
      - description: 部门 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除部门
      tags:
      - 组织管理
    get:
      description: 按 ID 读取当前租户下的部门
      parameters:
      - description: 部门 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取部门详情
      tags:
      - 组织管理
    put:
      consumes:
      - application/json
      description: 更新当前租户下的部门；只覆盖请求中出现的非空字段
      parameters:
      - description: 部门 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 部门信息（name、code、description、managerId、parentId）
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新部门
      tags:
      - 组织管理
  /api/v1/org/departments/tree:
    get:
      description: 返回当前租户的部门层级树
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取部门树
      tags:
      - 组织管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"